refactor: split host bulk routes
